# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PairEval API
  version: 0.1.0-M1
  license:
    name: MIT
    identifier: MIT
  description: |
    API ของระบบประเมินผลงานนักศึกษาแบบ pairwise comparison

    **ขอบเขตของเอกสารนี้คือ M1 Walking Skeleton** — ทุก endpoint ที่นี่สืบกลับไปหา
    user story ใน `docs/backlog.md` ได้ และทุก story ใน Sprint 1 มี endpoint รองรับ

    หลักที่ยึดทั้งไฟล์
    - ทุก endpoint ยกเว้น `/api/health` ต้องมี bearer token
    - ตรวจสิทธิ์ที่ server เสมอ (FR-AUTHZ-01) — ห้ามพึ่ง UI ในการซ่อน
    - ขอ resource ของ classroom ที่ไม่ได้สังกัด ตอบ **404 ไม่ใช่ 403**
      เพื่อไม่ให้รู้ว่า resource นั้นมีอยู่จริง (FR-AUTHZ-02)
    - error ใช้รูปแบบเดียวกันทั้งระบบ และ `code` คงที่สำหรับเครื่องอ่าน (FR-API-03)
    - timestamp ทุกตัวเป็น UTC (DR-05)
servers:
  - url: https://paireval-api.onrender.com
    description: Staging
  - url: http://localhost:8000
    description: Local
security:
  - bearerAuth: []
...
paths:
...
  /api/assignments:
    post:
      tags: [assignments]
      summary: สร้างงานประเมิน
      description: |
        อ้าง US-04 — งานที่สร้างใหม่มีสถานะ `DRAFT` เสมอ
        ผลรวมน้ำหนักของ criteria แต่ละ side จะถูกตรวจตอน publish ไม่ใช่ตอนสร้าง
      operationId: createAssignment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignmentCreate'
      responses:
        '201':
          description: สร้างสำเร็จ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assignment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationFailed'
...
  /api/assignments/{assignmentId}:publish:
    post:
      tags: [assignments]
      summary: เผยแพร่งานและสร้างคู่ประเมินทั้งหมด
      description: |
        อ้าง US-06 — สร้าง pair ทั้งหมดเก็บลง database ตอนนี้ ไม่สุ่มตอน runtime (FR-PAIR-01)
        ให้ `pairing_seed` เดิมแล้วผลต้องเหมือนเดิมทุกครั้ง (FR-PAIR-09)
      operationId: publishAssignment
      parameters:
        - $ref: '#/components/parameters/AssignmentId'
      responses:
        '200':
          description: เผยแพร่สำเร็จ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: 'สถานะไม่ใช่ DRAFT จึง publish ไม่ได้ (`code: INVALID_STATE`)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: 'ผลรวมน้ำหนัก criteria ไม่เท่ากับ 100 (`code: WEIGHT_SUM_INVALID`)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /api/assignments/{assignmentId}:recompute:
    post:
      tags: [scoring]
      summary: คำนวณคะแนนชั่วคราวใหม่
      description: |
        อ้าง US-13, US-16 — คำนวณด้วย `is_final=false` ทับ interim เดิมเสมอ เรียกได้ตลอด
        หลัง publish ไม่ต้องรอ deadline (FR-SCORE-06) ใช้ดูความคืบหน้าระหว่างทาง
        ไม่เปลี่ยนสถานะ assignment และไม่ถูก audit เพราะยังไม่ใช่คะแนนที่เชื่อถือได้
      operationId: recomputeScores
      parameters:
        - $ref: '#/components/parameters/AssignmentId'
      responses:
        '200':
          description: คำนวณเสร็จ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecomputeResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: 'role ไม่มี capability `MANAGE_ASSIGNMENT` (`code: FORBIDDEN`)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/assignments/{assignmentId}:finalize:
    post:
      tags: [scoring]
      summary: ตัดสินและประกาศคะแนน
      description: |
        อ้าง US-13 — เฉพาะ OWNER (capability `FINALIZE_SCORES`) เรียกได้ ต้องเลย
        `groupDeadlineUtc` ไปแล้วและ assignment อยู่ในสถานะ `PUBLISHED`/`OPEN`/`CLOSED`
        คำนวณ `is_final=true` เป็น append-only batch ใหม่ และเก็บ `score_snapshot` ที่มี
        input/output ครบชุด — ตรวจย้อนหลังได้แม้ reopen/finalize ใหม่หรือสูตรเปลี่ยนแล้ว

        ถ้ามี item ที่ผู้ประเมินยังไม่ถึงเกณฑ์ขั้นต่ำ (`LOW_CONFIDENCE`) endpoint นี้จะปฏิเสธ
        ด้วย 422 จนกว่าจะส่ง `confirmLowConfidence: true` มายืนยัน — ระบบไม่ finalize ให้อัตโนมัติ
      operationId: finalizeAssignment
      parameters:
        - $ref: '#/components/parameters/AssignmentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FinalizeAssignment'
      responses:
        '200':
          description: ประกาศผลแล้ว
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinalizeResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: 'role ไม่มี capability `FINALIZE_SCORES` — มีเฉพาะ OWNER (`code: FORBIDDEN`)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: 'สถานะไม่ใช่ PUBLISHED/OPEN/CLOSED หรือยังไม่ถึง deadline (`code: ALREADY_EXISTS`)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: 'มี item ติด LOW_CONFIDENCE และยังไม่ได้ส่ง confirmLowConfidence (`code: VALIDATION_FAILED`)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/assignments/{assignmentId}:reopen:
    post:
      tags: [scoring]
      summary: เปิดคะแนนที่ finalize แล้วกลับมาแก้
      description: |
        อ้าง US-13 — เฉพาะ OWNER เรียกได้ ต้องเป็นงานที่สถานะ `FINALIZED` เท่านั้น
        สถานะหลัง reopen กลับไปเป็น `CLOSED` (ไม่ใช่ `PUBLISHED`) ต้องเรียก `:finalize` ใหม่
        อีกครั้งเพื่อประกาศผล ต้องระบุเหตุผลเสมอเพราะกระทบความเชื่อมั่นของนักศึกษาที่เห็นคะแนน
        ไปแล้ว (`AuditAction.SCORES_REOPENED` อยู่ใน `ACTIONS_REQUIRING_REASON`)
      operationId: reopenAssignment
      parameters:
        - $ref: '#/components/parameters/AssignmentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReopenAssignment'
      responses:
        '200':
          description: เปิดกลับมาแก้แล้ว
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    enum: [CLOSED]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: 'role ไม่มี capability `FINALIZE_SCORES` (`code: FORBIDDEN`)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: 'ยังไม่ได้ finalize (`code: ALREADY_EXISTS`)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: 'ไม่ได้ระบุเหตุผล (`code: VALIDATION_FAILED`)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/assignments/{assignmentId}/comparisons:export:
    get:
      tags: [scoring]
      summary: ส่งออก raw comparison เป็น CSV (ปกปิดตัวตนผู้ประเมิน)
      description: |
        อ้าง US-15, FR-EXPORT-03 — ผู้เรียกที่มี capability `MANAGE_ASSIGNMENT` เรียกได้
        (OWNER และ CO_TEACHER) เพราะไฟล์นี้**ไม่มี**ตัวตนจริงของผู้ประเมินอยู่เลย มีแต่รหัส
        `E1, E2, ...` ที่สุ่มลำดับใหม่ทุกครั้งที่เรียก — ใช้ข้ามการ export ครั้งอื่นไม่ได้

        UTF-8 พร้อม BOM (FR-EXPORT-01) กันข้อความไทยเพี้ยนตอนเปิดด้วย Excel
        เอาเฉพาะ comparison ที่ `SUBMITTED` แล้ว (S2) — DRAFT ไม่เคยเข้าเงื่อนไขการคำนวณ
        คะแนน จึงไม่ควรอยู่ในไฟล์ที่ใช้ตรวจสอบคะแนนย้อนหลังเช่นกัน

        ต้องการตัวตนจริงของผู้ประเมิน ใช้ `POST .../comparisons:export-identified` แทน
      operationId: exportComparisons
      parameters:
        - $ref: '#/components/parameters/AssignmentId'
      responses:
        '200':
          description: ไฟล์ CSV
          content:
            text/csv:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/assignments/{assignmentId}/comparisons:export-identified:
    post:
      tags: [scoring]
      summary: ส่งออก raw comparison พร้อมอีเมลจริงของผู้ประเมิน
      description: |
        อ้าง US-15, FR-EXPORT-04 — เฉพาะ OWNER (capability `VIEW_EVALUATOR_IDENTITY`)
        ต้องระบุเหตุผลเสมอ (คือการ "ยืนยันเจตนา" ตาม AC) และถูกบันทึกลง audit log
        ทุกครั้งแบบไม่มีข้อยกเว้น (`AuditAction.IDENTIFIED_EXPORT`) — เรียกสำเร็จแล้วไม่มี
        ร่องรอยเป็นไปไม่ได้ เพราะ audit เขียนใน transaction เดียวกับการ export (FR-AUDIT-01)
      operationId: exportComparisonsIdentified
      parameters:
        - $ref: '#/components/parameters/AssignmentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
                  minLength: 1
      responses:
        '200':
          description: ไฟล์ CSV พร้อมอีเมลจริงของผู้ประเมิน
          content:
            text/csv:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: 'role ไม่มี capability `VIEW_EVALUATOR_IDENTITY` — มีเฉพาะ OWNER (`code: FORBIDDEN`)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          description: 'ไม่ได้ระบุเหตุผล (`code: VALIDATION_FAILED`)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/assignments/{assignmentId}/score-overrides:
    post:
      tags: [scoring]
      summary: แก้คะแนนที่คำนวณได้ด้วยมือ
      description: |
        อ้าง US-13, FR-SCORE-08 — เฉพาะ OWNER เรียกได้ บันทึกเป็นแถวใหม่เสมอ (ประวัติสะสม
        ไม่ทับของเดิม) ค่าที่มีผลจริงคือแถวล่าสุดของ item/เกณฑ์นั้น ต้องระบุเหตุผลเสมอ
        และถูกบันทึกลง audit log ครบ actor/ก่อน/หลัง/เหตุผล/เวลา (FR-AUDIT-02)
      operationId: createScoreOverride
      parameters:
        - $ref: '#/components/parameters/AssignmentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScoreOverrideCreate'
      responses:
        '201':
          description: บันทึกการ override แล้ว
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScoreOverrideResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: 'role ไม่มี capability `FINALIZE_SCORES` (`code: FORBIDDEN`)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          description: 'ไม่ได้ระบุเหตุผล (`code: VALIDATION_FAILED`)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/assignments/{assignmentId}/my-score:
    get:
      tags: [evaluation]
      summary: คะแนนของฉัน
      description: |
        อ้าง US-10 — สมาชิกห้องเรียนคนไหนก็เรียกได้ (ไม่ต้องมี capability พิเศษ)
        ถ้า assignment ยังไม่ `FINALIZED` ตอบ 200 พร้อม `finalized: false` และ `message`
        อธิบาย — **ไม่มีตัวเลขคะแนนใด ๆ ติดมาเลย** ไม่ใช่ 404 เพราะ "ยังไม่ประกาศ" เป็นสถานะ
        ปกติของ workflow ไม่ใช่ error

        คะแนนรายบุคคล (`individualComponent`) จะเป็น `null` และ `individualHidden: true`
        ถ้าจำนวนผู้ประเมินรายบุคคลของ item นี้ยังไม่ถึงเกณฑ์ k-anonymity (US-15 —
        ใช้ `min_comparisons` ของ assignment เป็นเกณฑ์ร่วม ไม่มีคอลัมน์แยก) เมื่อซ่อน
        `finalScore` (ที่รวมคะแนนรายบุคคลไว้ด้วย) จะถูกซ่อนไปด้วยเช่นกัน ไม่ใช่แค่
        `individualComponent` เพียงอย่างเดียว — เพื่อไม่ให้เดาย้อนกลับได้จากผลรวม
      operationId: getMyScore
      parameters:
        - $ref: '#/components/parameters/AssignmentId'
      responses:
        '200':
          description: คะแนนของผู้เรียก (หรือสถานะ "ยังไม่ประกาศผล" ถ้ายังไม่ finalize)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyScore'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: token ที่ได้หลัง login ผ่าน Google OIDC (US-01)
...
  responses:
    Unauthorized:
      description: 'ไม่ได้แนบ token หรือ token หมดอายุ (`code: UNAUTHENTICATED`)'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: 'login แล้วแต่ role ไม่มีสิทธิ์ทำสิ่งนี้ (`code: FORBIDDEN`)'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: |
        ไม่พบ resource — **รวมถึงกรณีที่ resource มีอยู่จริงแต่อยู่คนละ classroom**
        ตั้งใจตอบ 404 แทน 403 เพื่อไม่ให้รู้ว่ามีอยู่จริง (FR-AUTHZ-02)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
...
    AssignmentCreate:
      type: object
      required:
        - classroomId
        - name
        - groupMaxScore
        - individualMaxScore
        - groupDeadlineUtc
        - criteria
      properties:
        classroomId:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
        description:
          type: string
        artifactUrl:
          type: string
          format: uri
        groupMaxScore:
          type: number
          minimum: 0
        individualMaxScore:
          type: number
          minimum: 0
          description: ใส่ 0 แปลว่าไม่มีการประเมินรายบุคคล ระบบจะไม่สร้าง pair ฝั่งนั้น
        groupDeadlineUtc:
          type: string
          format: date-time
        individualDeadlineUtc:
          type: [string, 'null']
          format: date-time
        targetCoverage:
          type: integer
          minimum: 1
          maximum: 20
          default: 5
        maxWorkload:
          type: integer
          minimum: 1
          maximum: 30
          default: 8
        criteria:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/CriterionCreate'
    Assignment:
      type: object
      required: [id, classroomId, name, status, criteria]
      properties:
        id:
          type: string
          format: uuid
        classroomId:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
          enum: [DRAFT, PUBLISHED, OPEN, CLOSED, FINALIZED, ARCHIVED]
        groupDeadlineUtc:
          type: string
          format: date-time
        individualDeadlineUtc:
          type: [string, 'null']
          format: date-time
        criteria:
          type: array
          items:
            $ref: '#/components/schemas/Criterion'
...